+++ src/decorators/componentRouteAuthorized.decorator.ts
@@ -1,9 +1,9 @@
 import {ComponentRoute, ComponentRouteDefinition} from '@anglr/common/router';
 
-import {AuthGuard} from './auth.guard';
+import {authGuard} from '../guards';
 
 /**
- * Defines route for component on which is this decorator applied, automatically adds `AuthGuard`
+ * Defines route for component on which is this decorator applied, automatically adds `authGuard`
  * @param route - Definition of route
  */
 export function ComponentRouteAuthorized(route: ComponentRouteDefinition): ClassDecorator
@@ -11,9 +11,9 @@ export function ComponentRouteAuthorized(route: ComponentRouteDefinition): Class
     route.canActivate ??= [];
 
     //adds auth guard if there is no one
-    if(route.canActivate.indexOf(AuthGuard) < 0)
+    if(route.canActivate.indexOf(authGuard) < 0)
     {
-        route.canActivate.push(AuthGuard);
+        route.canActivate.push(authGuard);
     }
 
     return ComponentRoute(route);