+++ src/directives/index.ts
@@ -1,0 +1,2 @@
+export * from './authorize/authorize.directive';
+export * from './letAuthorized/letAuthorized.directive';