+++ src/services/index.ts
@@ -1,0 +1,2 @@
+export * from './authentication/authentication.service';
+export * from './authentication/authenticationServiceOptions.interface';